requirements.txt
pandas
faiss-cpu
numpy
sentence-transformers
requests
